+++ .env.example
@@ -2,3 +2,5 @@
 KAKAO_CLIENT_ID=
 # 카카오 앱의 REST API 키
 KAKAO_REDIRECT_URI=
+#Jwt secret key
+JWT_SECRET=
