+++ src/main/java/gift/kakao/controller/KakaoOrderController.java
@@ -2,7 +2,7 @@ package gift.kakao.controller;
 
 import gift.kakao.dto.OrderRequestDto;
 import gift.kakao.dto.OrderResponseDto;
-import gift.kakao.service.KakaoService;
+import gift.kakao.service.OrderService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,10 +14,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/orders")
 public class KakaoOrderController {
-    private final KakaoService kakaoService;
+    private final OrderService orderService;
 
-    public KakaoOrderController(KakaoService kakaoService) {
-        this.kakaoService = kakaoService;
+    public KakaoOrderController(OrderService orderService) {
+        this.orderService = orderService;
     }
 
     @PostMapping
@@ -27,7 +27,7 @@ public class KakaoOrderController {
     ) {
         Long memberId = (Long) request.getAttribute("memberId");
 
-        OrderResponseDto responseDto = kakaoService.order(memberId, requestDto);
+        OrderResponseDto responseDto = orderService.order(memberId, requestDto);
 
         return ResponseEntity.status(HttpStatus.CREATED)
             .body(responseDto);
