--- src/main/java/gift/kakao/dto/KakaoLoginResponseDto.java
@@ -1,15 +1,0 @@
-package gift.kakao.dto;
-
-import gift.member.entity.Member;
-
-public record KakaoLoginResponseDto(
-    String token,
-    Member member
-) {
-    public static KakaoLoginResponseDto from(String token, Member member) {
-        return new KakaoLoginResponseDto(
-            token,
-            member
-        );
-    }
-}
