+++ src/main/java/gift/kakao/template/Content.java
@@ -4,4 +4,5 @@ package gift.kakao.template;
 public record Content(
     String title,
     Link link
-) {}
+) {
+}
