+++ src/main/java/gift/member/entity/SocialType.java
@@ -1,0 +1,6 @@
+package gift.member.entity;
+
+public enum SocialType {
+    KAKAO,
+    ETC
+}
