+++ src/main/java/gift/member/repository/MemberRepository.java
@@ -1,11 +1,12 @@
 package gift.member.repository;
 
 import gift.member.entity.Member;
+import gift.member.entity.SocialType;
 import gift.member.vo.Email;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Member findByEmail(Email email);
-    Optional<Member> findByProviderAndSocialId(String provider, Long socialId);
+    Optional<Member> findByProviderAndSocialId(SocialType provider, Long socialId);
 }
