+++ src/main/java/gift/member/service/MemberService.java
@@ -1,6 +1,5 @@
 package gift.member.service;
 
-import gift.kakao.dto.KakaoLoginResponseDto;
 import gift.kakao.dto.KakaoUserResponseDto;
 import gift.member.dto.MemberLoginRequestDto;
 import gift.member.dto.MemberLoginResponseDto;
@@ -15,5 +14,5 @@ public interface MemberService {
     MemberResponseDto findById(Long id);
     void update(Long id, MemberRegisterRequestDto requestDto);
     void delete(Long id);
-    KakaoLoginResponseDto loginWithKakao(KakaoUserResponseDto userResponseDto);
+    MemberLoginResponseDto loginWithKakao(String code);
 }
