+++ src/main/java/gift/option/service/OptionService.java
@@ -10,5 +10,4 @@ public interface OptionService {
     OptionResponseDto findByProductIdAndOptionId(Long productId, Long optionId);
     OptionResponseDto update(Long productId, Long optionId, OptionRequestDto requestDto);
     void delete(Long productId, Long optionId);
-    void subtract(Long optionId, Integer num);
 }
