+++ src/main/resources/application.properties
@@ -15,7 +15,7 @@ spring.h2.console.path=/h2-console
 spring.sql.init.mode=always
 
 # jwt secret-key
-jwt.secret=Yn2kjibddFAWtnPJ2AFlL8WXmohJMCvigQggaEypa5E=
+jwt.secret=${JWT_SECRET}
 jwt.token-validity-time=3600
 
 # jpa
